pyproject.toml
[project]
name = "genai-accelerator-labs"
version = "0.1.0"
description = "rag"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.1",
    "numpy>=2.2.5",
    "nylas>=6.9.0",
    "openai>=1.77.0",
    "pydantic-ai>=0.2.3",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "tiktoken>=0.1.1",
    "fastapi>=0.115.12",
    "sqlalchemy>=2.0.41",
    "celery>=5.5.2",
    "redis>=6.1.0",
    "pgvector>=0.4.1",
    "psycopg[binary]>=3.2.9",
    "docling>=2.15.1",
    "langfuse>=2.60.7",
    "sentry-sdk>=2.29.1",
    "logfire>=3.16.1",
    "llm-guard>=0.3.16",
    "pip>=25.1.1",
    "datasets>=3.6.0",
    "seaborn>=0.13.2",
    "mcp[cli]>=1.8.1",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
